refactor(deterministic): extract security checks into a mixin
skills/evaluators/deterministic.py was 1510 LOC dominated by 11
security-focused check methods (~415 LOC) and 6 regex pattern lists
(SUDO, EXTERNAL_NETWORK, _SECRET, _EXFIL, _DESTRUCTIVE, _INJECTION),
mixed in among the file/tool/command/output check methods.

Extracted into:
- _security_patterns.py — all 6 pattern lists as module constants (90 LOC)
- _security_checks.py — SecurityChecksMixin with all 11 security
  check methods, importing patterns from _security_patterns (380 LOC)

deterministic.py is now 1086 LOC, with DeterministicEvaluator
inheriting from SecurityChecksMixin. The class still has no __init__
and no shared state, so the mixin is purely a physical-organization
mechanism — runtime behavior is byte-identical.

SUDO_PATTERNS and EXTERNAL_NETWORK_PATTERNS re-exported from
deterministic.py with noqa: F401 to preserve any external imports.

Zero functional change. 1792 tests pass (same baseline). Including
all 96 tests in tests/skills/test_security.py.
